% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMultiHetSep.R
\name{generateMultiHetSep}
\alias{generateMultiHetSep}
\title{Write out multiHetSep (input of MSMC/PSMC') of simulated individual}
\usage{
generateMultiHetSep(msprimeVCFPath, outpath)
}
\arguments{
\item{outpath}{Output path for the multiHetSep file that is being generated.}

\item{msprimeVCF}{Path to the msprime generated VCF.}
}
\value{
Returns nothing. But writes a multihetsep (input to PSMC'/MSMC) file to the given outpath.
}
\description{
Write out multiHetSep (input of MSMC/PSMC') of simulated individual.
}
\examples{
msmcInference<-readMSMCInference(pathOfMSMCOutFinal = "../data/msprimeMultiHetSep/simulatedMsprime.oak.msmc.out.final.txt", mutationRate = 1e-8)
populationParametersChange<-generatePopulationParameterChanges(msmcInference)
generateMSPrimeFunction( outPath = "../code/msmc2msprime_Feb1.py", PopulationParametersChangeInput=populationParametersChange)
source_python("../code/msmc2msprime_Feb1.py")
simulation<-msmc_model(length=3e5, seed=30, mu=1.01e-08)
writeVCFOfSimulation(simulation, outpath = "../data/msprimeQLob.vcf", ploidy=2)
generateMultiHetSep(msprimeVCF = "../data/msprimeQLob.vcf",outpath = "../data/msprimeQLob_multihetsep.txt")
}
