% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePopulationParameterChanges.R
\name{generatePopulationParameterChanges}
\alias{generatePopulationParameterChanges}
\title{Generate the population parameter changes for msprime.}
\usage{
generatePopulationParameterChanges(msmcInference)
}
\arguments{
\item{msmcInference}{The tibble that readMSMCInference generates.}
}
\value{
A character that desribes the population parameter changes msprime will simulate. Based entirely off the msmcInference.
}
\description{
Uses msmcInference tibble to generate the population parameter changes for a msprime function.
}
\examples{
msmcInference<-readMSMCInference(pathOfMSMCOutFinal = "../data/msprimeMultiHetSep/simulatedMsprime.oak.msmc.out.final.txt", mutationRate = 1e-8)
populationParametersChange<-generatePopulationParameterChanges(msmcInference)
}
