% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMSMCInference.R
\name{readMSMCInference}
\alias{readMSMCInference}
\title{Read in the final ouput of MSMC.}
\usage{
readMSMCInference(pathOfMSMCOutFinal, mutationRate)
}
\arguments{
\item{pathOfMSMCOutFinal}{The path to the "out.final.txt" that MSMC outputs.}

\item{mutationRate}{The mutation rate in mutations/bp/gen. This is the same mutation rate that one would use to scale results to real time and population sizes. See https://github.com/stschiff/msmc/blob/master/guide.md for more help.}
}
\value{
A tibble desribing the scaled population size changes over generations.
}
\description{
Reads in the out.final.txt file from MSMC and converts the output into a tibble.
}
\examples{
readMSMCInference(pathOfMSMCOutFinal = "../data/msprimeMultiHetSep/simulatedMsprime.oak.msmc.out.final.txt", mutationRate = 1e-8)

}
