% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeVCFOfSimulation.R
\name{writeVCFOfSimulation}
\alias{writeVCFOfSimulation}
\title{Write out VCF of simulated individual}
\usage{
writeVCFOfSimulation(simulation, outpath, ploidy = 2)
}
\arguments{
\item{simulation}{An <msprime.trees.TreeSequence> object}

\item{outpath}{The desired output path for the VCF you are trying to create.}

\item{ploidy}{An integer that defines the ploidy of the individual you simulated. The default is 2, referring to diploid.}
}
\value{
Returns nothing. But writes a VCF file to the given outpath
}
\description{
Uses msprime and the reticulate package to write out a VCF to the simulated indidividual
}
\examples{
msmcInference<-readMSMCInference(pathOfMSMCOutFinal = "../data/msprimeMultiHetSep/simulatedMsprime.oak.msmc.out.final.txt", mutationRate = 1e-8)
populationParametersChange<-generatePopulationParameterChanges(msmcInference)
generateMSPrimeFunction( outPath = "../code/msmc2msprime_Feb1.py", PopulationParametersChangeInput=populationParametersChange)
source_python("../code/msmc2msprime_Feb1.py")
simulation<-msmc_model(length=3e5, seed=30, mu=1.01e-08)
writeVCFOfSimulation(simulation, outpath = "../data/msprimeQLob.vcf", ploidy=2)
}
