% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMSPrimeFunction.R
\name{generateMSPrimeFunction}
\alias{generateMSPrimeFunction}
\title{Generate the msprime function describing the inferred demography from MSMC}
\usage{
generateMSPrimeFunction(outPath, PopulationParametersChangeInput)
}
\arguments{
\item{outPath}{The desired output path for the msprime script you are trying to create.}

\item{PopulationParametersChangeInput}{The character that defines the how the population sizes change across generations. This is generated by the generatePopulationParameterChanges.}
}
\value{
Returns nothing. But writes to file an msprime function called "msmc_model" that describes the inference made by MSMC
}
\description{
Generates the  msprime function describing the inferred demography from MSMC using the populationParameterChanges
}
\examples{
msmcInference<-readMSMCInference(pathOfMSMCOutFinal = "../data/msprimeMultiHetSep/simulatedMsprime.oak.msmc.out.final.txt", mutationRate = 1e-8)
populationParametersChange<-generatePopulationParameterChanges(msmcInference)
generateMSPrimeFunction( outPath = "../code/msmc2msprime_Feb1.py", PopulationParametersChangeInput=populationParametersChange)
}
